Add standalone build infrastructure for reftable
The reftable library is integrates with Git, so it become a first class
supported storage mechanism, but is sufficiently complex that other
projects (e.g. libgit2) may want to consume the same source code.

To make this possible, the library also compiles completely standalone, which is
demonstrated with the Bazel BUILD/WORKSPACE files that this commit adds.

Signed-off-by: Han-Wen Nienhuys <hanwen@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>